+++ eulerpi/core/sampling.py
@@ -19,11 +19,11 @@ import emcee
 import numpy as np
 from schwimmbad import MultiPool
 
-from epipy import logger
-from epipy.core.kde import calc_kernel_width
-from epipy.core.model import Model
-from epipy.core.result_manager import ResultManager
-from epipy.core.transformations import eval_log_transformed_density
+from eulerpi import logger
+from eulerpi.core.kde import calc_kernel_width
+from eulerpi.core.model import Model
+from eulerpi.core.result_manager import ResultManager
+from eulerpi.core.transformations import eval_log_transformed_density
 
 # TODO: This works on the blob
 # Return the samples.