+++ eulerpi/examples/cpp/python_reference_plants.py
@@ -2,7 +2,7 @@ import jax.numpy as jnp
 import numpy as np
 from jax import jacrev, jit
 
-from epipy.core.model import ArtificialModelInterface, JaxModel, Model
+from eulerpi.core.model import ArtificialModelInterface, JaxModel, Model
 
 
 class JaxPlant(JaxModel, ArtificialModelInterface):