+++ eulerpi/examples/simple_models.py
@@ -3,7 +3,7 @@ from typing import Optional
 import jax.numpy as jnp
 import numpy as np
 
-from epipy.core.model import ArtificialModelInterface, JaxModel
+from eulerpi.core.model import ArtificialModelInterface, JaxModel
 
 
 class Linear(JaxModel, ArtificialModelInterface):