+++ CHANGELOG.md
@@ -17,6 +17,18 @@ All notable changes to this project will be documented in this file.
 
 - Bug in module B that caused a crash
 
+## [0.1.5] - 2022-04-06
+
+### Changed
+
+- Renamed epipy to eulerpi
+
+## [0.1.4] - 2022-04-06
+
+### Changed
+
+- Updated installation instructions
+
 ## [0.1.3] - 2022-04-05
 
 ### Changed